#8: Create navigations between pages - Added 3 simple(base) views; - Added BaseViewModel class; - Added NotifyPropertychanged class; - Changed structure of the project;
