+++ View/Childs/AddNewWordView.xaml
@@ -1,0 +1,18 @@
+﻿<UserControl
+    x:Class="LearnAndGrow.View.Childs.AddNewWordView"
+    xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+    xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
+    xmlns:local="clr-namespace:LearnAndGrow.View"
+    xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+    d:DesignHeight="450"
+    d:DesignWidth="800"
+    mc:Ignorable="d">
+    <Grid>
+        <TextBlock
+            HorizontalAlignment="Center"
+            VerticalAlignment="Center"
+            FontSize="46"
+            Text="Add New Word" />
+    </Grid>
+</UserControl>
