+++ View/Childs/LearnNewWordsView.xaml
@@ -1,16 +1,14 @@
-﻿<Window
-    x:Class="LearnAndGrow.LearnAndGrowView"
+﻿<UserControl
+    x:Class="LearnAndGrow.View.Childs.LearnNewWordsView"
     xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
     xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
-    xmlns:local="clr-namespace:LearnAndGrow"
+    xmlns:local="clr-namespace:LearnAndGrow.View.Childs"
     xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-    xmlns:models="clr-namespace:LearnAndGrow.Model"
-    xmlns:viewModels="clr-namespace:LearnAndGrow.ViewModel"
-    Title="LearnAndGrowView"
-    Width="800"
-    Height="450"
-    d:DataContext="{d:DesignInstance Type=viewModels:LearnAndGrowViewModel}"
+    xmlns:viewModels="clr-namespace:LearnAndGrow.ViewModel.Childs"
+    d:DataContext="{d:DesignInstance Type=viewModels:LearnNewWordsViewModel}"
+    d:DesignHeight="450"
+    d:DesignWidth="800"
     mc:Ignorable="d">
     <Grid>
         <Grid.RowDefinitions>
@@ -75,4 +73,4 @@
                 FontSize="46" />
         </Grid>
     </Grid>
-</Window>
+</UserControl>
