+++ View/LearnAndGrowWindow.xaml.cs
