+++ ViewModel/Childs/AddNewWordViewModel.cs
@@ -1,0 +1,13 @@
+﻿using LearnAndGrow.Utilities;
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace LearnAndGrow.ViewModel.Childs
+{
+    internal class AddNewWordViewModel : BaseViewModel
+    {
+    }
+}
