+++ ViewModel/Childs/WelcomeViewModel.cs
@@ -1,0 +1,15 @@
+﻿using LearnAndGrow.Utilities;
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace LearnAndGrow.ViewModel.Childs
+{
+    public class WelcomeViewModel : BaseViewModel
+    {
+        public WelcomeViewModel()
+        { }
+    }
+}
